Moved test_forward to intgration tests
